### 2021-01-25 #2987 @danielskatz
The previous version of this paper (which was not reviewed as being too small) is https://github.com/openjournals/joss-reviews/issues/2758
