### 2021-02-09 73b4348
Fix matplotlib text margin
