### 2021-02-19 #2987 @xuanxu
@ygrange, @astrom-tom would you be willing to review this paper? I think your knowledge of Astrophysics and Python makes both of you a great fit for this submission.
