### 2021-02-19 #2987 @ygrange
Looks cool (I don't really have a radio telescope at hand to test it against though). But I'd be happy to have a look at it :)
