### 2021-02-19 #2987 @0xCoto
@ygrange - Thank you very much for agreeing to review the paper. As it is understandably unlikely that a reviewer would have access to appropriate hardware, we have included an example observation file obtained with Virgo on the repository, for software testing and verification purposes: https://github.com/0xCoto/Virgo/tree/master/joss_example
