### 2021-02-26 #3067 @whedon
[REVIEW]: Virgo: A Versatile Spectrometer for Radio Astronomy
**Submitting author:** @0xCoto (<a href="http://orcid.org/0000-0001-6928-6877">Apostolos Spanakis-Misirlis</a>)
**Repository:** <a href="https://github.com/0xCoto/Virgo" target ="_blank">https://github.com/0xCoto/Virgo</a>
**Version:** v3.7.0
