### 2021-03-30 #3067 @ygrange
The goal of the tool is to make radio telescopes accessible to a larger community, mainly targeting amateurs, students and educators. I'm struggling a bit with the target audience. This is because I personally think this is a very great group to focus on, because the field of radio astronomy tends to have a bit of an image to be hard to start working in, and those are important groups in general. On the other hand I can imagine the JOSS target audience to be more scientists than educators or amateurs. Would it be possible to add a sentence about use by professional astronomers, even though they may not necessarily be the original target audience?
