### 2021-03-30 #3067 @astrom-tom
Here is the first part of the review. I will create a few issues in the repository. I will continue the review once these problem have been addressed.

General comment:
For such a software I believe that having the complete documentation inside the readme
is probably not appropriate and restrain a bit its content.
For python development, I would advise to use Sphinx to create a documentation that can
be then 'attached' directly to Github. It gives a great flexibility on the content
and its organization. That would greatly help the readibility. A nice example can be seen
here https://fpavogt.github.io/fcmaker/. You could as well incorporate all the comments below.

A statement of need:
The statement of need is not present in the readme (even if it is in the paper)

Installation instructions:
There is something that is not clear from the installation part of the README. It is stated:
'To use Virgo, make sure Python and GNU Radio (with gr-osmosdr) are installed on your machine.'

Is 'gr-osmosdr' part of GNU Radio? When I go to the gr-osmosdr website it is said
'Marcus D. Leech has kindly integrated the forementioned build steps into his gnuradio installation script at "This is the most user-friendly option so far.'

So it is not clear what I have to install. I could install gnu-radio without problem but I will not install other things if they are not necessary. Please make this
clear.

Also, you mention, in 'Usage' to use 'python virgo.py [arguments]'. I believe that if I install the package via pip I do not
have access directly to virgo.py unless I go to the installation directory and use it from here (which is not the ma)

Example usage:
The example usage of the readme is relatively short and could be expanded and better explained.

Functionality documentation:
This is an important problem. There is no documentation of the function that are coded in the package.
For example, let's take the function given in the example usage:

>>>import virgo
>>>help(virgo.predict)
Help on function predict in module virgo.virgo:
predict(lat, lon, height=0, source='', date='', plot_sun=True, plot_file='')

>>>help(virgo.observe)
Help on function observe in module virgo.virgo:
observe(obs_parameters, spectrometer='wola', obs_file='observation.dat', start_in=0)

--> There is no documentation of what it does, the input format and the output that is given.

Automated tests:
I did not see any automated test mention in the readme.

Community guidelines:
None are given in the readme.
