### 2021-03-30 0xCoto/Virgo#10 @astrom-tom
JOSS review: Documentation general comment
Dear @0xCoto.

Here is a global comment about the documentation. I will create other issues with specific comments.

For such a software I believe that having the complete documentation inside the readme
is probably not appropriate and restrain a bit its content.
For python development, I would advise to use Sphinx to create a documentation that can
be then 'attached' directly to Github. It gives a great flexibility on the content
and its organization. That would greatly help the readibility. A nice example can be seen
here https://fpavogt.github.io/fcmaker/.

If you want to keep the readme format, you will have to at least add the statement of need (it is in the paper) and the community guidelines
