### 2021-03-30 0xCoto/Virgo#11 @astrom-tom
JOSS review: Installation and Usage
Dear @0xCoto

Here are some comments about the instruction:

There is something that is not clear from the installation part of the README. It is stated:
'To use Virgo, make sure Python and GNU Radio (with gr-osmosdr) are installed on your machine.'

Is 'gr-osmosdr' part of GNU Radio? When I go to the gr-osmosdr website it is said
'Marcus D. Leech has kindly integrated the forementioned build steps into his gnuradio installation script at "This is the most user-friendly option so far.'

So it is not clear what I have to install. I could install gnu-radio without problem but I will not install other things if they are not necessary. Please make this clearer.

Usage:
Also, you mention, in 'Usage' to use 'python virgo.py [arguments]'. I believe that if I install the package via pip I do not
have access directly to virgo.py unless I go to the installation directory and use it from here (which is not the point of installing a package)

Example Usage:
The example usage of the readme is relatively short and could be expanded and better explained. What are you trying to do, and what do you expect as a result?
