### 2021-03-30 0xCoto/Virgo#12 @astrom-tom
JOSS review: functionality documentation & testing
Hi @0xCoto

**Functionality documentation:**

The _in-code_ documentation is an important problem.
There is no documentation of the functions that are coded in the package.
For example, let's take the function given in the example usage:

```
>>>import virgo
>>>help(virgo.predict)
Help on function predict in module virgo.virgo:
predict(lat, lon, height=0, source='', date='', plot_sun=True, plot_file='')`

>>>help(virgo.observe)
Help on function observe in module virgo.virgo:
observe(obs_parameters, spectrometer='wola', obs_file='observation.dat', start_in=0)
```

--> There is no documentation of what it does, the input format and the output that is given. Please document every function of the package.

**Testing:**
In addition, I did not see any automated test mention in the readme. You must be able to test your code. It might be an automatic test suite (with unittest or pytest package for example) or at least a  manual procedure that tell us how to test your code.
