### 2021-03-30 #3067 @astrom-tom
Dear @xuanxu,
I created the three following issues for the paper.
[JOSS review: Documentation general comment #10](https://github.com/0xCoto/Virgo/issues/10#issue-845188047)
[JOSS review: Installation and Usage #11 ](https://github.com/0xCoto/Virgo/issues/11#issue-845191559)
[JOSS review: functionality documentation & testing #12 ](https://github.com/0xCoto/Virgo/issues/12#issue-845208507)

There is a fair amount of work to be done to have the paper accepted. I am happy to give a hand where I can!
