+++ virgo/virgo.py
@@ -9,6 +9,17 @@ import shutil
 import warnings
 
 def simulate(l, b, beamwidth=0.6, v_min=-400, v_max=400, plot_file=''):
+	'''
+	Simulate 21 cm profiles based on the LAB HI Survey.
+	
+	Args:
+		l: float. Target galactic longitude [deg]
+		b: float. Target galactic latitude [deg]
+		beamwidth: float. Telescope half-power beamwidth (approx. equal to 0.7 * lambda/D) [deg]
+		v_min: float. Minimum radial velocity (xlim) [km/s]
+		v_max: float. Maximum radial velocity (xlim) [km/s]
+		plot_file: string. Output plot filename
+	'''
 	import requests
 	import matplotlib
 	import matplotlib.pyplot as plt
@@ -96,6 +107,18 @@ def simulate(l, b, beamwidth=0.6, v_min=-400, v_max=400, plot_file=''):
 	plt.close()
 
 def predict(lat, lon, height=0, source='', date='', plot_sun=True, plot_file=''):
+	'''
+	Plots source Alt/Az given the observer's Earth coordinates.
+	
+	Args:
+		lat: float. Observer latitude [deg]
+		long: float. Obesrver longitude [deg]
+		height: float. Observer elevation [m]
+		source: string. Source name
+		date: string. Date in YYYY-MM-DD format. If no date is given, it defaults to today's system date.
+		plot_sun: bool. Also plot Sun position for reference
+		plot_file: string. Output plot filename
+	'''
 	from astropy.time import Time
 	from astropy.visualization import astropy_mpl_style, quantity_support
 	from astropy.coordinates import SkyCoord, EarthLocation, AltAz, get_sun
@@ -186,6 +209,16 @@ def predict(lat, lon, height=0, source='', date='', plot_sun=True, plot_file='')
 	plt.clf()
 
 def equatorial(alt, az, lat, lon, height=0):
+	'''
+	Takes observer's location and Alt/Az as input and returns RA/Dec as a tuple.
+	
+	Args:
+		alt: float. Altitude [deg]
+		az: float. Azimuth [deg]
+		lat: float. Observer latitude [deg]
+		lon: float. Observer longitude [deg]
+		height: float. Observer elevation [m]
+	'''
 	from astropy.time import Time
 	from astropy.coordinates import SkyCoord, EarthLocation, AltAz
 	import astropy.units as u
@@ -210,6 +243,13 @@ def equatorial(alt, az, lat, lon, height=0):
 	return (ra, dec)
 
 def galactic(ra, dec):
+	'''
+	Converts RA/Dec. to galactic coordinates, returning galactic longitude and latitude (tuple).
+	
+	Args:
+		ra: float. Right ascension [deg]
+		dec: float. Declination [deg]
+	'''
 	from astropy.time import Time
 	from astropy.coordinates import SkyCoord, EarthLocation, AltAz
 	import astropy.units as u
@@ -225,6 +265,12 @@ def galactic(ra, dec):
 	return (l, b)
 
 def frequency(wavelength):
+	'''
+	Transform wavelength to frequency.
+	
+	Args:
+		wavelength: float. Wavelength [m]
+	'''
 	# Define speed of light
 	c = 299792458.0
 
@@ -234,6 +280,12 @@ def frequency(wavelength):
 	return f
 
 def wavelength(frequency):
+	'''
+	Transform frequency to wavelength.
+	
+	Args:
+		frequency: float. Wave frequency [Hz]
+	'''
 	# Define speed of light
 	c = 299792458.0
 
@@ -243,6 +295,15 @@ def wavelength(frequency):
 	return l
 
 def gain(D, f, e=0.7, u='dBi'):
+	'''
[…]
