### 2021-04-01 c9ebe77
Gain -> Aperture fix
