### 2021-04-01 4e3134e
Add CLI install
