### 2021-04-01 85b2e2b
Change script to command
