+++ README.md
@@ -65,101 +65,7 @@ A list of GR-supported SDRs can be found [here](https://wiki.gnuradio.org/index.
 - - Radiometer equation (S:N estimation)
 - Built-in tool for conducting rapid RFI surveys
 - Argument-parsing support
-- Works directly via `python virgo.py`, or as a module (see below)
-
-## Example Usage
-
-```python
-import virgo
-
-# Define observation parameters
-obs = {
-    'dev_args': '',
-    'rf_gain': 10,
-    'if_gain': 20,
-    'bb_gain': 20,
-    'frequency': 1420e6,
-    'bandwidth': 5e6,
-    'channels': 2048,
-    't_sample': 1,
-    'duration': 60
-}
-
-# Check source position
-virgo.predict(lat=39.83, lon=-74.87, source='Cas A', date='2020-12-26')
-
-# Begin data acquisition in 10 sec
-virgo.observe(obs_parameters=obs, obs_file='observation.dat', start_in=10)
-
-# Analyze data, mitigate RFI and export the data as a FITS file
-virgo.plot(obs_parameters=obs, n=20, m=35, f_rest=1420.4057517667e6,
-           obs_file='observation.dat', cal_file='calibration.dat',
-           rfi=[1419.2e6, 1419.3e6], waterfall_fits='obs.fits',
-           slope_correction=True, plot_file='plot.png')
-```
-
----
-
-### Function definitions
-```python
-# Schedules observation for start_in seconds. Spectrometer defaults to WOLA unless 'ftf' is specified.
-observe(obs_parameters, spectrometer='wola', obs_file='observation.dat', start_in=0)
-
-# Plots data. n and m are median in the spectrum and time series respectively, f_rest is used for frequency -> velocity transformation, slope_correction used to correct poor spectra (linear regression), dB to display data values in decibels, rfi to mask out contaminated data, xlim and ylim to scale frequency and time range respectively, obs_file and cal_file are the data files of the observation and calibration respectively, waterfall_fits is the output .fits filename and spectra_csv and power_cst are the output .csv files respectively.
-plot(obs_parameters='', n=0, m=0, f_rest=0, slope_correction=False, dB=False, rfi=[0,0], xlim=[0,0], ylim=[0,0], dm=0,
-	 obs_file='observation.dat', cal_file='', waterfall_fits='', spectra_csv='', power_csv='', plot_file='plot.png'):
-
-# Plots source AltAz given observer's Earth coordinates. If date is not given, it defaults to today's system date.
-predict(lat, lon, height=0, source='', date='', plot_sun=True, plot_file='')
-
-# Simulates 21 cm profiles based on the LAB HI Survey. l and b are galactic longitude and latitude, beamwidth is the HPBW (beam FWHM) of the telescope and v_min/v_max is the x axis velocity limits.
-simulate(l, b, beamwidth=0.6, v_min=-400, v_max=400, plot_file=''):
-
-# Converts RA/Dec. to galactic longitude and latitude. The position is returned as a tuple, where the first element [0] is l and the second [1] is b.
-galactic(ra, dec)
-
-# Takes observer's location and AltAz as input and returns RA/Dec as a tuple.
-equatorial(alt, az, lat, lon, height=0)
-
-# Converts frequency (Hz) to wavelength (m)
-frequency(wavelength)
-
-# Converts wavelength (m) to frequency (Hz)
-wavelength(frequency)
-
-# Computes the gain of a parabolic antenna. D is diameter (m), f is frequency (Hz), e is aperture efficiency and u is the output unit (dBi, linear or K/Jy)
-gain(D, f, e=0.7, u='dBi')
-
-# Computes the effective antenna aperture (m^2). Gain is in dBi, f is in Hz.
-A_e(gain, f)
-
[…]
