### 2021-04-02 9f3a256
Clarify CLI usage
