### 2021-04-02 0xCoto/Virgo#10 @0xCoto
@astrom-tom - Thank you for raising this important issue and suggesting a solution.

The README has just been slightly shortened (subtracting specific details concerning the usage of the package), and the documentation has been composed and can be found **[here](https://virgo.readthedocs.io/en/latest/)**.

Since it is unlikely that a reviewer would have access to appropriate hardware, we have included an example observation file obtained with Virgo on the [repository](https://github.com/0xCoto/Virgo/tree/master/joss_example) for software testing and verification purposes, as described [here](https://virgo.readthedocs.io/en/latest/examples.html#offline-experiments). Please let me know if this is sufficient, or if you were envisioning something else.

Lastly, the contributor guidelines have been composed and added to the [README](https://github.com/0xCoto/Virgo/blob/master/README.md#contributing) along with the statement of need.
