### 2021-04-02 0xCoto/Virgo#11 @0xCoto
The installation instructions have been rewritten to be more explicitly clear about the role of the `gr-osmosdr` block. Since GNU Radio is used for the acquisition and early-stage (real-time) digital signal processing of the data, unless you _do_ happen to have an SDR at hand, I wouldn't consider it a necessity for testing the package.

All of the built-in functionalities (other than `virgo.observe`) can be tested by simply running `pip install astro-virgo` and do not require GNU Radio/`gr-osmosdr` or any particular hardware. To avoid future confusions, this clarification has been added to the [documentation](https://virgo.readthedocs.io/en/latest/installation.html#dependencies).

As for the usage, you are correct that it's not obvious how one would run the script with command-line arguments. We've just pushed an update for this, and the user can now run Virgo from the command line with e.g. `virgo -h` (although as described in the documentation, this method is typically preferred for quick hardware-verification tests etc. rather than extended functionality, so I'd suggest experimenting with Virgo as a module for maximum versatility). There are no particular instructions for getting the command built into one's system, as [`setup.py`](https://github.com/0xCoto/Virgo/blob/master/setup.py) is properly configured to automatically install the command upon the installation of the package (e.g. with `pip`).

The example usage has been transferred to the [documentation](https://virgo.readthedocs.io/en/latest/examples.html) with an explanation for each function called. The [reference page](https://virgo.readthedocs.io/en/latest/reference.html) furthermore includes function-specific clarifications for each parameter.
