### 2021-04-02 0xCoto/Virgo#12 @0xCoto
In-code documentation has been implemented for every function.

For example,
```python
>>> import virgo
>>> help(virgo.predict)
```
now clarifies the role of the function and the properties of each parameter:
```
predict(lat, lon, height=0, source='', date='', plot_sun=True, plot_file='')
    Plots source Alt/Az given the observer's Earth coordinates.

    Args:
            lat: float. Observer latitude [deg]
            lon: float. Obesrver longitude [deg]
            height: float. Observer elevation [m]
            source: string. Source name
            date: string. Date in YYYY-MM-DD format. If no date is given, it defaults to today's system date.
            plot_sun: bool. Also plot Sun position for reference
            plot_file: string. Output plot filename
```
---
Since it is unlikely that a reviewer would have access to appropriate hardware, we have included an example observation file obtained with Virgo on the [repository](https://github.com/0xCoto/Virgo/tree/master/joss_example) for software testing and verification purposes, as described [here](https://virgo.readthedocs.io/en/latest/examples.html#offline-experiments). Please let me know if this is sufficient, or if you were envisioning something else.
