### 2021-04-02 6375427
Statement of need addition & change script to CLI
