### 2021-04-02 #3067 @0xCoto
@ygrange - Thank you very much for your comments.

> I noticed that there are four authors on the paper, while only one did contribute to the repo. This is not against the JOSS policy. They are also explicitly thanked for their contributions in the README. Just wanted to make sure you are aware of the relevant policy on co-athorship.

The paper only contains three authors (I assume by "four" you were referring to the GitHub README), which have had major contributions to the project (not only in terms of code, but also theoretical design of certain technical aspects), and to the composition of the paper as well. Only three of the people listed in the README footer have had a sufficient level of contribution to the development of the package and were considered for co-authorship.

The lack of commits by co-authors is because communication (sharing & reviewing code, as well as theoretical development) has been achieved via direct-messaging platforms, so many commits pushed by me are essentially the result of significant help provided by the co-authors.

---

> You cite the FITS 3.0 standard in the paper. I think you basically import from astropy.io.fits and I couldn't really find what version that is. Since version 4.0 seems to have recently appeared I was wondering whether you have a reason to cite this specific version, or that this is more of a "generic FITS" citation.

This citation was indeed intended more as a generic FITS reference, but we can also cite version 4.0 instead if that seems more appropriate.

---

> Seems like the DOI proposed by Whedon for the book is wrong btw.

Yes, the book does not have a DOI, so this 'warning' should be dismissed.

---

> The goal of the tool is to make radio telescopes accessible to a larger community, mainly targeting amateurs, students and educators. I'm struggling a bit with the target audience. This is because I personally think this is a very great group to focus on, because the field of radio astronomy tends to have a bit of an image to be hard to start working in, and those are important groups in general. On the other hand I can imagine the JOSS target audience to be more scientists than educators or amateurs. Would it be possible to add a sentence about use by professional astronomers, even though they may not necessarily be the original target audience?

I've appended a sentence to the Statement of Need. I'd like to note that Virgo is being used for undergraduate experiments in universities around the world, and I'm aware of a student heavily basing his thesis on the package to complete his degree.
