### 2021-04-02 #3067 @0xCoto
@astrom-tom - Thank you for the comments raised, they have all been addressed in the corresponding issues and the software definitely benefits from the requested additions.

@ygrange - Since you mentioned continuation issues, I'll also inform you of the documentation that has just been composed and can be found [here](https://virgo.readthedocs.io/en/latest/). Hope that helps; please let me know if anything is unclear, and I'd be more than happy to clarify.
