### 2021-04-02 #3067 @ygrange
Thanks for that! And thanks for all the answers. Especially the last comment is one I struggle with to make, because I by no means want to discredit the use of education/outreach as part of science (quite the opposite)!
I will have a look at the docs, install instructions and follow them and continue writing comments (probably going to be after the easter break though).
