### 2021-05-14 0xCoto/Virgo#11 @ygrange
Cool to read the install docs! I would advice you to add something to the installation section in the README where you discuss the installation to add a sentence like the Note in the docs, with a link to the docs for more info.
