### 2021-05-14 0xCoto/Virgo#11 @ygrange
Also: Is installing the github master version a good idea for users? If so it may be useful to add to the docs (seems as easy as clonign the repo and run `pip install .` ).
