### 2021-05-14 0xCoto/Virgo#12 @ygrange
Let me reply to this one too :):

docstrings: Great addition!

I think having unit tests (which would be testing every function for every type of input/output, but of course there are very pragmatic ways to go about this) can help you catch subtle issues (e.g. input generating wrong output, or error not being caught, etc) better than having a single test for the whole, but in general I would say that giving users who just installed the tool a way to prove that the install on their system runs as epected is a good start (and I think that is pretty much what you did).

Does the manual test you define cover the whole functionality in the code? I will run it and if I have feedback on them, you'll hear it :).
