### 2021-05-14 0xCoto/Virgo#10 @ygrange
I do not really understand point 1 in the contribution section. For the rest: good that it's there :)
