### 2021-05-14 #3067 @ygrange
One of mu check boxes is " State of the field: Do the authors describe how this software compares to other commonly-used packages?" I have too little knowledge of this specific field to give some examples, but I can imagine there are several tools which may have different target audiences. Are you aware of other tools that have similar goals as Virgo?
