+++ docs/contributing.md
@@ -7,11 +7,9 @@ Please note the Code of Conduct. We hope you follow it in all your interactions 
 
 ## Pull Requests
 
-1. Ensure any install or build dependencies are removed before the end of the layer when doing a 
-   build.
-2. Update the README.md with details of changes to the interface, this includes new environment 
+1. Update the README.md with details of changes to the interface, this includes new environment 
    variables, exposed ports, useful file locations and container parameters.
-3. Increase the version numbers in any examples files and the README.md to the new version that this
+2. Increase the version numbers in any examples files and the README.md to the new version that this
    Pull Request would represent. The versioning scheme Virgo uses is [SemVer](http://semver.org/).
 
 ## Code of Conduct
