### 2021-05-14 0xCoto/Virgo#10 @0xCoto
@ygrange - Thanks for pointing that out. I've taken it out (a52df12) after a second read because it doesn't really make much sense for this package indeed.
