### 2021-05-14 9f057f3
Extend examples/tests
