### 2021-05-14 0xCoto/Virgo#12 @0xCoto
@ygrange - I just extended the script so that it is more complete in terms of tests/examples (a384417).
