### 2021-05-14 32e7c87
Add installation note from docs
