### 2021-05-14 0xCoto/Virgo#12 @ygrange
Yeah, that's a nice way to show the install was successful (you may want to consider change it from joss_example to something like sanity_check (or something more appropriate) but I totally leave that to you. You may want to mention near the installation instructions that you can use this example to test the installation (and at the same time it is a nice example of functionality; I've just downloaded some Dwingeloo telescope data from CAMRAS' public page to test with and I am using this one as entry point :) ; I can tell you I do get a (partial) plot).
