### 2021-05-14 0xCoto/Virgo#12 @ygrange
The thing between parentheses brings me to the next question: is it sensible to want to use Virgo to download some public data and process it with the plot function? Because in that case it would be nice to have a small section in the docs that tells you what properties the data should have. I'm sure the data I am using to play with is not very appropriate (I noticed when trying to manually debug a bit that it seems to contain nans that do not get caught, but anyhow I get a bunch of numpy warnings on inappropriate numbers further on and I do not know whether that would be because of the telescope settings, or that I somehow should set certain features on and off.
(I hope you get the point I am trying to make here :) ).
