### 2021-05-14 0xCoto/Virgo#12 @0xCoto
Yeah, I was just looking through the public CAMRAS data archive because I was curious how you were planning to go about this. Ideally, you want to use the `.plot` function to visualize data generated with Virgo's `.observe` function, because the format is pretty specific. I know that the Dwingeloo telescope is equipped with a USRP SDR device (which is perfectly compatible and has been extensively tested with Virgo), but the data found on the archive is for the most part post-processed in all sorts of unique ways, so it would be a bit tricky (though not impossible) to parse some raw data and view them with Virgo, although I don't know how useful that would be.

For example, I could run the exact same code on an 18-meter telescope equipped with the same USRP model and it would be a test of similar quality I think, since the interface is basically identical (only the antenna aperture and RF frontend is different of course, but that isn't really relevant to such tests). If you'd like to avoid the hassle of observing with Dwingeloo etc., I could quickly acquire and share an example `18m.dat` any time for you to test `.plot` against if that helps. You are of course welcome to provide any custom parameters for the `.observe` function by the way!
