### 2021-05-14 0xCoto/Virgo#11 @0xCoto
> I would advice you to add something to the installation section in the README where you discuss the installation to add a sentence like the Note in the docs, with a link to the docs for more info.

Good idea, added!

> Is installing the github master version a good idea for users?

Yup, that's perfectly reasonable - nothing stopping you from installing the package this way.

Just added a mention for this possibility to the [installation section](https://virgo.readthedocs.io/en/latest/installation.html).
