### 2021-05-16 #3067 @0xCoto
Regarding other packages, I've seen a few home-built independent efforts by newcomers writing up their own software for basic observations (just averaging spectra, for the most part), but they are nowhere near as close to the ease of use, installation, cross-platform support, functionality and documentation that this package has to offer. While marginally functional, there have been very few released solutions for public use like plugins for other software (limited to data acquisition, which is only a small fraction of what a typical user needs), but they are extremely limited compared to Virgo's versatility, and are not commonly used, nor maintained.
