### 2021-05-16 0xCoto/Virgo#12 @ygrange
I guessed that CAMRAS had somehow been tested by looking at the author list :). I tried it with some satellite observations (basically because those seemed raw data to me) from here https://charon.camras.nl/public/satellites/ (selection mostly based on which link does not give me a 404 error. That gives me one plot that actually contains data (and a bunch of errors which probably have to do with NaNs in the data).

Yeah, I'd love to have a .dat set, but I assume it won't be too different from the example already in the repo? Would having some sort of boundary conditions to data on which the plot function would work (I realise this is much simpler said than done) so that as a user I could judge whether I can use some public data or not ("observed with Virgo" is of course a nice zeroeth order boundary condition).
