### 2021-05-16 0xCoto/Virgo#12 @0xCoto
Just to clarify, during observing, Virgo doesn't spit out raw I/Q data (like the ones found on the CAMRAS archive you linked), and here's why: raw I/Q data files contain **all** acquired information about the recorded signal (pre-FFT/discretized time-domain data points). The problem is that this information can get pretty large at high sample rates, long integration durations and ADC sample depths (bits), which can obviously get very impractical in the real world. The important thing to note, however, is that for many/most observations, raw I/Q information is not at all necessary.

Virgo tackles this problem by running the FFT (WOLA or plain computationally-inexpensive Fourier transform filterbank), channelizing and partially integrating the acquired data on the fly (in real time). This drastically reduces the data size from sometimes TBs to only a few MBs or GBs of information, containing nothing more than the user needs.

Therefore, the output .dat binary file produced by the package only contains FFT samples, and when processed with Numpy, this gets expanded to a 2D array (where each column represents the FFT bin/channel and each row represents the corresponding partially-averaged time bin). `virgo.plot` then takes care of processing and analyzing the data to produce each plot and apply all sorts of computations to the data given the user's input.

That being said, I'll look into building a GNU Radio-based I/Q to .dat parser. This would allow you to take the raw data from any instrument and view, process and analyze the data with Virgo however you wish (as long as the essential observation parameters/headers are known of course).
