### 2021-05-17 0xCoto/Virgo#12 @ygrange
> That being said, I'll look into building a GNU Radio-based I/Q to .dat parser. This would allow you to take the raw data from any instrument and view, process and analyze the data with Virgo however you wish (as long as the essential observation parameters/headers are known of course).
Yeah, that sounds cool. I think this would be a great addition. With my hat of reviewer on, I want to add that in principle I don't consider it as being required to be part of the reviewed version, but more as a general idea for extension post-review.
