### 2021-05-17 #3067 @ygrange
I think with this state of the field, one of the extra added values is reproducibility. If everybody uses the same package, it makes the results much better to interpret and reproduce IMO. I personally would advice you to add the first sentence (in some rewording) to the introduction or statement of need section, since I think it is a pretty strong argument (to me at least).

That's really the only point left from me. If you agree with the comment, we're totally good to go (and if you don't I'm sure we can find a good solution here :) ).

Two broader points of advice that I consider future work:
- For future versions I would advice you to have a look at more granular unit testing.
- Maybe it could be a good idea to extend the documentation with some sort of tutorial for new users who have a radio telescope lying around.
- There is another point in one of the issues: would be neat if it would be even more generally usable to use public data from (amateur) radio facilities.
