### 2021-05-19 0xCoto/Virgo#10 @astrom-tom
Hey @0xCoto ,

Thanks for the  new documentation! It looks good! One detail that would be nice is to have the table of content on the left as well.
If you can do that I will close this ticket!

Thanks!
