### 2021-05-19 0xCoto/Virgo#10 @astrom-tom
Great! Did you put the community guidelines somewhere because I can not find them.
