### 2021-05-19 0xCoto/Virgo#10 @0xCoto
They are in [this readme section](https://github.com/0xCoto/Virgo#contributing).

https://github.com/0xCoto/Virgo/blob/master/docs/contributing.md
