### 2021-05-19 0xCoto/Virgo#12 @astrom-tom
Hey!
So I went through the documentation about the functions [here](https://virgo.readthedocs.io/en/latest/reference.html). I am happy to see that the functions are documented. Nevertheless, you always document the parameters of the function but not the output.  For example. the beamwidth function surely has an output. This should be properly documented.

I will comment on the testing in another comment
