### 2021-05-19 b9487af
Add output type
