### 2021-05-19 0xCoto/Virgo#11 @astrom-tom
That all seems very good! Thank for updating everything in the documentation. The dependencies are now much clearer and the Usage and Example are also good. Great work!
