### 2021-05-19 0xCoto/Virgo#12 @astrom-tom
> Just added the output type for all functions. Hope it's a bit more obvious now.

is it in the website documentation as well?
