### 2021-05-20 #3067 @0xCoto
Just wanted to inform that the last sentence of the first paragraph of the Statement of Need section has been reworded a bit to give a slightly more fair representation for the state of other work:

> Although in-house acquisition solutions have been independently created by many, composing their own software for simple observations, these are not comparable in terms of installation count, ease of use, cross-platform support, functionality, reproducibility, maintenance and level of documentation that this package has to offer.

@ygrange - I don't expect many objections, but wanted to let you know of the change so you're aware as a reviewer.
