### 2021-05-20 0xCoto/Virgo#12 @astrom-tom
Oh ok. I think it would be clearer to have, below the 'Parameters' list, the list of output (in the same format at input). Also, as you precise the unit of some of the parameters (that's awesome), you should precise the unit of the output to be complete.
