### 2021-05-20 0xCoto/Virgo#12 @0xCoto
Could you please clarify what you mean by 'list of output'? What would this look like for something like the beamwidth function for example which only returns one value?

As for the output units, that's a good point. Will add them.
