### 2021-05-20 0xCoto/Virgo#12 @astrom-tom
![Screenshot_20210520_140131](https://user-images.githubusercontent.com/26314073/119027198-0cfa3080-b974-11eb-9b69-2704b7302363.png)

Let's take this example: You have the list of parameters format like

Parameters: -S
                    -sefd
                    -t
                    -bw
...

It would be nice to have also the list of ouputs in the exact same format:

output: - SNR, signal to noise ratio, unitless, float
