### 2021-05-20 0xCoto/Virgo#12 @ygrange
I like the new wording �

--
Y. Grange

> Op 14 mei 2021 om 18:03 heeft Apostolos ***@***.***> het
volgende geschreven:
>
> ﻿
> @ygrange - I just extended the script so that it is more complete in
terms of tests/examples (a384417).
>
