### 2021-05-20 0xCoto/Virgo#12 @0xCoto
@astrom-tom Is something like [this](https://virgo.readthedocs.io/en/latest/reference.html) sufficient?

![Screen Shot 2021-05-20 at 9 54 15 PM](https://user-images.githubusercontent.com/25392776/119033466-ec52ca80-b9b5-11eb-9dcc-e672ccaf4a62.png)
![Screen Shot 2021-05-20 at 9 54 00 PM](https://user-images.githubusercontent.com/25392776/119033440-e5c45300-b9b5-11eb-9aca-61642d6c9e8f.png)
