### 2021-05-20 0xCoto/Virgo#12 @astrom-tom
For the testing part, I understood that the test you proposed is to make sure that the installation was done correctly without error. This is great but I am not sure this is what JOSS is referring to when they talk about testing. I will check with the editor and come back to you.
