### 2021-05-31 #3067 @0xCoto
@xuanxu - Version number: [v3.7.0](https://github.com/0xCoto/Virgo/tree/v3.7.0)

[Zenodo](https://zenodo.org/record/4778382#) DOI: 10.5281/zenodo.4778382
