### 2021-06-05 #3067 @Kevin-Mattheus-Moerman
@0xCoto I've proofread your paper and it looks good so I'll now proceed to accept this work in JOSS. Congratulations. Thanks @astrom-tom and @ygrange for reviewing this work!
